requirements.txt
pandas
faiss-cpu
sentence-transformers
transformers
jupyter
notebook
numpy
pytest
